feat: 로그아웃 모달 생성 및 불필요한 로직 삭제
