+++ src/pages/My/components/profiles/Setting.tsx
@@ -1,19 +1,10 @@
-import { useNavigate } from 'react-router';
 import CImg from '../../../../components/common/CImg';
 import CBreadcrumb from '../../../../components/common/CBreadcrumb';
 import { right } from '../../../../assets/index';
-import { useAuthStore } from '../../../../store/useAuthStore';
-import { useUserStore } from '../../../../store/useUserStore';
-import { logoutAPI } from '../../../../api/authApi';
-import { useQueryClient } from '@tanstack/react-query';
 import { useModalStore } from '../../../../store/useModalStore';
 
 const Setting = () => {
-  const navigate = useNavigate();
   const openMoadl = useModalStore((state) => state.openModal);
-  const queryClient = useQueryClient();
-  const { refreshToken, clearToken } = useAuthStore();
-  const setUserInfo = useUserStore((state) => state.setUserInfo);
 
   const serviceItems = [
     { id: 'terms', label: '이용약관', url: 'https://www.notion.so/34f61e1109d08034b2b5d21a227baf4e?source=copy_link' },
@@ -27,20 +18,6 @@ const Setting = () => {
     { id: 'withdraw', label: '회원 탈퇴' },
   ];
 
-  const logoutClickHandler = async () => {
-    const res = await logoutAPI(refreshToken);
-
-    if (res.code === 200) {
-      clearToken();
-      setUserInfo({ userId: null, name: '', email: '', profileImageUrl: '' });
-
-      queryClient.removeQueries({ queryKey: ['userInfo'] });
-      alert('로그아웃 되었습니다!');
-
-      navigate('/home');
-    }
-  };
-
   const goExternalLink = (e: React.MouseEvent, url: string | null) => {
     e.stopPropagation(); // 중요: 화살표 클릭 시 체크박스가 토글되는 것을 막아줍니다!
     if (url) {
@@ -50,7 +27,7 @@ const Setting = () => {
 
   const handleClick = (id: string) => {
     if (id === 'logout') {
-      logoutClickHandler();
+      openMoadl('LOGOUT');
     } else if (id === 'withdraw') {
       openMoadl('WITHDRAW');
     }
